Consolidate ProtocolLane into shared header file
